Alternative start

+++ src/main/java/org/pbl/expression/Number.java
@@ -2,14 +2,14 @@ package org.pbl.expression;
 
 public class Number extends Expression {
 
-    public int num;
+    public Double num;
 
-    public Number(int value) {
+    public Number(double value) {
         this.num = value;
     }
 
     @Override
     public String toString() {
-        return Integer.toString(num);
+        return Double.toString(num);
     }
 }
